cc-wrapper: Use response files for cc
This works around issues on OS that has a limited on the number of
arguments that can be passed to a process.

This change will passes all arguments in a temporary response file,
but only if we were able to expand response params with
expandResponseParams (that way we can safely assume there are
no unexpanded response params).

It also treats -L and -l linker options passed to cc-wrapper.sh
specially.  It puts these in a response file and adds that file to
the list of options with `-Wl,@...`.  This helps avoid issues when
the wrapped cc binary itself calls the linker.

Finally because ld cannot handle response files it shrinks the -L
options before passing them to ld itself using temporary local
symlinks with short names.